+++ src/lib/custom-typings/gulp-mocha.d.ts
@@ -1,5 +1,0 @@
-declare module "gulp-mocha" {
-  type StaticGulpMocha = (options?: any) => any;
-  const staticGulpMocha: StaticGulpMocha;
-  export = staticGulpMocha;
-}