+++ src/lib/custom-typings/gulp-pug.d.ts
@@ -1,5 +1,0 @@
-declare module "gulp-pug" {
-  type StaticGulpPug = (...args: any[]) => any;
-  const staticGulpPug: StaticGulpPug;
-  export = staticGulpPug;
-}