+++ src/lib/custom-typings/gulp-typedoc.d.ts
@@ -1,5 +1,0 @@
-declare module "gulp-typedoc" {
-  type StaticGulpTypedoc = (...args: any[]) => any;
-  const staticGulpTypedoc: StaticGulpTypedoc;
-  export = staticGulpTypedoc;
-}