+++ src/lib/custom-typings/merge2.d.ts
@@ -1,5 +1,0 @@
-declare module "merge2" {
-  type StaticMerge2 = (streams: any[]) => any;
-  const staticMerge2: StaticMerge2;
-  export = staticMerge2;
-}